+++ Console/Assets/BlueConsole/Runtime/ConsoleTypeParameters.cs
@@ -1,64 +1,64 @@
 using System.Globalization;
 using UnityEngine;
 
-public class ConsoleTypeParameters : MonoBehaviour
+public static class ConsoleTypeParameters
 {
     [TypeParameter(@"[0|1]")]
-    public bool BoolParameter(string inputParam)
+    public static bool BoolParameter(string inputParam)
     {
         return inputParam == "1";
     }
 
     [TypeParameter(@"\d+")]
-    public byte ByteParameter(string inputParam)
+    public static byte ByteParameter(string inputParam)
     {
         return byte.Parse(inputParam);
     }
 
     [TypeParameter(@"[-+]?\d+")]
-    public sbyte SByteParameter(string inputParam)
+    public static sbyte SByteParameter(string inputParam)
     {
         return sbyte.Parse(inputParam);
     }
 
     [TypeParameter(@"[-+]?\d+")]
-    public int IntParameter(string inputParam)
+    public static int IntParameter(string inputParam)
     {
         return int.Parse(inputParam);
     }
 
     [TypeParameter(@"[-+]?\d+")]
-    public uint UIntParameter(string inputParam)
+    public static uint UIntParameter(string inputParam)
     {
         return uint.Parse(inputParam);
     }
 
     [TypeParameter(@"[-+]?([0-9]*[.])?[0-9]+([eE][-+]?\d+)?")]
-    public decimal DecimalParameter(string inputParam)
+    public static decimal DecimalParameter(string inputParam)
     {
         return decimal.Parse(inputParam, NumberStyles.Any, CultureInfo.InvariantCulture);
     }
 
     [TypeParameter(@"[-+]?([0-9]*[.])?[0-9]+([eE][-+]?\d+)?")]
-    public double DoubleParameter(string inputParam)
+    public static double DoubleParameter(string inputParam)
     {
         return double.Parse(inputParam, NumberStyles.Any, CultureInfo.InvariantCulture);
     }
 
     [TypeParameter(@"[-+]?([0-9]*[.])?[0-9]+([eE][-+]?\d+)?")]
-    public float FloatParameter(string inputParam)
+    public static float FloatParameter(string inputParam)
     {
         return float.Parse(inputParam, NumberStyles.Any, CultureInfo.InvariantCulture);
     }
 
     [TypeParameter("\".*?\"", true)]
-    public string StringParameter(string inputParam)
+    public static string StringParameter(string inputParam)
     {
         return inputParam.Replace("\"", string.Empty);
     }
 
     [TypeParameter("'.'")]
-    public char CharParameter(string inputParam)
+    public static char CharParameter(string inputParam)
     {
         return inputParam.Replace("'", string.Empty).ToCharArray()[0];
     }