+++ Console/Assets/BlueConsole/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## [1.1.1] - 20.11.2023
+## [1.2.0] - 23.11.2023
  - Added input history recalling
  - Added hint accepting
  - Added a check for EventSystem in ConsoleController
+ - Added command parameters color
+ - Added GUI scaling
  - Improved commands parameters formatting
+ - Made all of command attributes optional
  - Fixed "ArgumentException" from regex matching while typing in console input
 
 ## [1.1.0] - 19.11.2023